+++ src/aqua_diagnostics/global_biases/global_biases.py
@@ -271,22 +271,25 @@ class GlobalBiases:
         if plev_max is None:
             plev_max = bias['plev'].max().item()
 
-        bias = bias.sel(plev=slice(plev_max, plev_min))
+        if plev_min < plev_max:
+            bias = bias.sel(plev=slice(plev_min, plev_max))
+        else:
+            bias = bias.sel(plev=slice(plev_max, plev_min))
 
         # Calculate the mean bias along the time axis
         mean_bias = bias.mean(dim='time')
         nlevels = 18
 
         # Calculate the zonal mean bias
-        zonal_bias = mean_bias.mean(dim='lon')
+        zonal_bias = mean_bias.mean(dim='lon').compute()
 
         # Determine colorbar limits if not provided
         if vmin is None or vmax is None:
-            vmin, vmax = zonal_bias.min(), zonal_bias.max()
+            vmin, vmax = zonal_bias.min().values, zonal_bias.max().values
             if vmin * vmax < 0:  # if vmin and vmax have different signs
                 vmax = max(abs(vmin), abs(vmax))
                 vmin = -vmax
-
+    
         levels = np.linspace(vmin, vmax, nlevels)
 
         title = (f"{var_name} vertical bias of {self.model} {self.exp} {self.startdate_data}/{self.enddate_data}\n"